chage int16

+++ bert/modeling.py
@@ -537,9 +537,9 @@ def create_attention_mask_from_input_mask(from_tensor, to_mask):
 
   to_shape = get_shape_list(to_mask, expected_rank=2)
   to_seq_length = to_shape[1]
-  #change here
+
   to_mask = tf.cast(
-      tf.reshape(to_mask, [batch_size, 1, to_seq_length]), tf.bfloat16)
+      tf.reshape(to_mask, [batch_size, 1, to_seq_length]), tf.float32)
 
   # We don't assume that `from_tensor` is a mask (although it could be). We
   # don't actually care if we attend *from* padding tokens (only *to* padding)
@@ -547,7 +547,7 @@ def create_attention_mask_from_input_mask(from_tensor, to_mask):
   #
   # `broadcast_ones` = [batch_size, from_seq_length, 1]
   broadcast_ones = tf.ones(
-      shape=[batch_size, from_seq_length, 1], dtype=tf.bfloat16)
+      shape=[batch_size, from_seq_length, 1], dtype=tf.float32)
 
   # Here we broadcast along two dimensions to create the mask.
   mask = broadcast_ones * to_mask
@@ -709,8 +709,7 @@ def attention_layer(from_tensor,
     # Since attention_mask is 1.0 for positions we want to attend and 0.0 for
     # masked positions, this operation will create a tensor which is 0.0 for
     # positions we want to attend and -10000.0 for masked positions.
-    #change here
-    adder = (1.0 - tf.cast(attention_mask, tf.bfloat16)) * -10000.0
+    adder = (1.0 - tf.cast(attention_mask, tf.float32)) * -10000.0
 
     # Since we are adding it to the raw scores before the softmax, this is
     # effectively the same as removing these entirely.
